requirements.txt
numpy==1.24.3
opencv-python==4.7.0.72
picamera2==0.3.9
opencv-contrib-python==4.7.0.72
Pillow==9.5.0
mysql-connector-python==8.3.0
RPi.GPIO==0.7.1
mfrc522==0.0.7
spidev==3.6

